requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
ansi2html==1.8.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.29.7
azure-identity==1.15.0
blinker==1.6.3
boto3==1.34.30
botocore==1.34.30
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.5
dash==2.14.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-mantine-components==0.12.1
dash-table==5.0.0
dash_daq==0.5.0
Deprecated==1.2.14
exceptiongroup==1.2.0
Flask==3.0.0
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
microsoft-kiota-abstractions==1.1.0
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.2.1
microsoft-kiota-serialization-json==1.0.1
microsoft-kiota-serialization-text==1.0.0
msal==1.24.1
msal-extensions==1.1.0
msgraph-core==1.0.0
msgraph-sdk==1.1.0
multidict==6.0.5
names==0.3.0
nest-asyncio==1.5.8
numpy==1.26.3
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
packaging==23.2
pandas==2.2.0
patsy==0.5.6
pendulum==3.0.0
plotly==5.18.0
plotly-express==0.4.1
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.4
s3transfer==0.10.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
statsmodels==0.14.1
std-uritemplate==0.0.50
tenacity==8.2.3
time-machine==2.13.0
typing_extensions==4.8.0
tzdata==2023.4
urllib3==1.26.18
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
